feat(app): EDITOR_KIT composition — markdown-editor becomes thin (WP2)
- basic marks/blocks/code/table/list/link kits gain their React halves
  (withComponent components relocated from markdown-editor's
  EDITOR_COMPONENTS map + v53 feature-owned autoformat inputRules:
  headings, quote, hr, fences, lists/todos, md-representable marks).
- markdown-editor.tsx is now just the seed/echo lifecycle around
  usePlateEditor({ plugins: EDITOR_KIT }); EDITOR_PLUGINS and
  EDITOR_COMPONENTS are deleted; kit-parity retargets to EDITOR_KIT.
- WP3's block-menu / floating-toolbar / block-placeholder kits land as
  typed empty arrays so the composition is final and WP3 file-disjoint.
- block-list.tsx splits its delegation affordance into todo-delegation.tsx
  behind React.lazy: eagerly importing vault-context from a kit-imported
  module closed an ESM cycle around base-kit (kit files → block-list →
  vault-context → markdown-doc → base-kit) and deadlocked module init
  when an entrypoint reached base-kit through a kit file first.
- mermaid fences render a lazy preview inside the code-block node with a
  Source/Diagram toggle (render-only; the node model never changes).
- inline-ai.ts → .tsx; the ai leaf's highlight styling moves onto the
  plugin via withComponent.
