+++ packages/app/src/editor/kits/block-placeholder-kit.tsx
@@ -1,0 +1,7 @@
+// Block placeholder kit — WP3 fills this file (BlockPlaceholderPlugin with
+// the per-type placeholder map; `query` must exclude `frontmatter`). Landed
+// empty in WP2 so EDITOR_KIT's composition is final.
+
+import type { SlatePlugins } from "platejs";
+
+export const BlockPlaceholderKit: SlatePlugins = [];
