+++ packages/app/src/editor/kits/floating-toolbar-kit.tsx
@@ -1,0 +1,8 @@
+// Floating toolbar kit — WP3 fills this file (the @platejs/floating swap of
+// selection-toolbar.tsx's positioning layer; the toolbar then renders from
+// this kit's `render.afterEditable` instead of markdown-editor.tsx JSX).
+// Landed empty in WP2 so EDITOR_KIT's composition is final.
+
+import type { SlatePlugins } from "platejs";
+
+export const FloatingToolbarKit: SlatePlugins = [];
